:zap: Move createErrorExecution to genericHelper
